moves menu items to ui
